+++ CHANGELOG.md
@@ -5,11 +5,18 @@
 ### Added
 
 - [TD-4438] Included-in-parameters reference dataset validation
+- [TD-5509]
+  - Data structure link labels graphql query
+  - Data structure link creation using source and target structure IDs
 - [TD-3806] Add cached content for implementations, rules, remediations plan and structures notes
 
 ### Changed
 
 - [TD-3879] All `raw keyword` indexing mappings uses empty string as null_value 
+
+### Fixed
+
+- [TD-5566] Remove template content from Remediation plan audit payload
 
 ## [5.2.1] 2023-03-08